SDI-2526-407-12-ejercicio-complementario-ejercicio1

+++ musicstoreapp/routes/api/songsAPIv1.0.js
@@ -1,4 +1,6 @@
 const {ObjectId} = require("mongodb");
+const {songValidatorInsert} = require("./songValidator");
+const {validationResult} = require("express-validator");
 
 module.exports = function (app, songsRepository, usersRepository) {
 
@@ -82,17 +84,52 @@ module.exports = function (app, songsRepository, usersRepository) {
         try {
             let songId = new ObjectId(req.params.id)
             let filter = {_id: songId}
-            songsRepository.deleteSong(filter, {}).then(result => {
-                if (result === null || result.deletedCount === 0) {
-                    res.status(404).json({error: "ID inválido o no existe, no se ha borrado el registro."});
+            isUserAuthenticated(req.user, songId).then(isAuthor => {
+                if (isAuthor) {
+                    songsRepository.deleteSong(filter, {}).then(result => {
+                        if (result === null || result.deletedCount === 0) {
+                            res.status(404).json({error: "ID inválido o no existe, no se ha borrado el registro."});
+                        } else {
+                            res.status(200).send(JSON.stringify(result));
+                        }
+                    }).catch(error => {
+                        res.status(500).json({error: "Se ha producido un error:" + error.message})
+                    });
                 } else {
-                    res.status(200).send(JSON.stringify(result));
-                }
-            }).catch(error => {
-                res.status(500).json({error: "Se ha producido un error:" + error.message})
-            });
+                res.status(401).json({error: "Usuario no autorizado a borrar la canción"})
+            }
+            })
         } catch (error) {
             res.status(500).json({error: error.message})
+        }
+    });
+
+    app.post('/api/v1.0/songs',songValidatorInsert, function (req, res) { //otra forma de hacerla
+        try {
+            const errors = validationResult(req);
+            if (!errors.isEmpty()) {
+                res.status(422);
+                res.json({errors: errors.array()});
+            }
+            let song = {
+                title: req.body.title,
+                kind: req.body.kind,
+                price: req.body.price,
+                author: req.res.user
+            }
+            songsRepository.insertSong(song, function (songId) {
+                if (songId === null) {
+                    res.status(409).json({error: "No se ha podido crear la canción. El recurso ya existe."});
+                } else {
+                    res.status(201).json({
+                        message: "Canción añadida correctamente.",
+                        _id: songId
+                    })
+                }
+            });
+        } catch (error) {
+            res.status(500).json({error: "Se ha producido un error al intentar crear la canción: " +
+                    error.message})
         }
     });
 
@@ -143,7 +180,7 @@ module.exports = function (app, songsRepository, usersRepository) {
      *                 error:
      *                   type: string
      */
-    app.post('/api/v1.0/songs', function (req, res) {
+    app.post('/api/v1.0/songs', function (req, res) { //hecha con funciones
         try {
             let song = {
                 title: req.body.title,
@@ -151,16 +188,23 @@ module.exports = function (app, songsRepository, usersRepository) {
                 price: req.body.price,
                 author: req.res.user
             }
-            songsRepository.insertSong(song, function (songId) {
-                if (songId === null) {
-                    res.status(409).json({error: "No se ha podido crear la canción. El recurso ya existe."});
-                } else {
-                    res.status(201).json({
-                        message: "Canción añadida correctamente.",
-                        _id: songId
-                    })
+            validatorInsertSong(song, function(errors){
+                if (errors !== null && errors.length > 0) {
+                    res.status(422);
+                    res.json({errors: errors});
+                } else{
+                    songsRepository.insertSong(song, function (songId) {
+                        if (songId === null) {
+                            res.status(409).json({error: "No se ha podido crear la canción. El recurso ya existe."});
+                        } else {
+                            res.status(201).json({
+                                message: "Canción añadida correctamente.",
+                                _id: songId
+                            })
+                        }
+                    });
                 }
-            });
+            })
         } catch (error) {
             res.status(500).json({error: "Se ha producido un error al intentar crear la canción: " +
                     error.message})
@@ -294,4 +338,56 @@ module.exports = function (app, songsRepository, usersRepository) {
         }
     });
 
+    function validatorInsertSong(song, callBackFunction){
+        let errors = [];
+
+        if (!song?.title?.trim())
+            errors.push({
+                "value": song.title,
+                "msg": "El título de la cancion no puede estar vacio",
+                "param": "title",
+                "location": "body"
+        })
+
+        if (!song?.kind?.trim())
+            errors.push({
+                "value": song.title,
+                "msg": "El género de la cancion no puede estar vacio",
+                "param": "kind",
+                "location": "body"
+            })
+
+        if (!song?.price)
+            errors.push({
+                "value": song.title,
+                "msg": "El precio de la cancion no puede estar vacio",
+                "param": "kind",
+                "location": "body"
+            })
+
+        if (!song?.price || Number(song.price) < 0)
+            errors.push({
+                "value": song.title,
+                "msg": "El precio de la cancion no puede ser negativo",
+                "param": "kind",
+                "location": "body"
+            })
+
+        if (errors.length <= 0)
+            callBackFunction(null);
+        else
+            callBackFunction(errors);
+
+    }
+
+    async function isUserAuthenticated(user, songId) {
+        let filterOptions = {$and: [{"_id": songId}, {"author": user}]};
+        let options = {};
+        try{
+            const songs = await songsRepository.getSongs(filterOptions, options)
+            return songs.length > 0;
+        } catch (error) {
+            return false;
+        }
+    }
 }
